docs: :memo: Revamp README.md with personal branding, updated bio, and enhanced project showcase 🚀
This commit enhances the README.md by incorporating a unique personal branding image, an updated bio that better reflects skills and experience, and a polished project showcase highlighting key achievements and technologies used. The goal is to provide a more professional, clear, and visually impactful view for visitors to the repository. 🌟

By @DanielSantiagoV